+++ gcovr/utils.py
@@ -217,6 +217,15 @@ class RelativeFilter(Filter):
 
     def match(self, path):
         abspath = os.path.realpath(path)
+
+        # On Windows, a relative path can never cross drive boundaries.
+        # If so, the relative filter cannot match.
+        if sys.platform == 'win32':
+            path_drive, _ = os.path.splitdrive(abspath)
+            root_drive, _ = os.path.splitdrive(os.path.realpath(self.root))
+            if path_drive != root_drive:
+                return None
+
         relpath = os.path.relpath(abspath, self.root)
         return super(RelativeFilter, self).match(relpath)